refactor(components): reorganize UI and Electron integration
- Replace HelloWorld component with a new Dashboard component
- Simplify and streamline ElectronIntegration component
- Move Electron integration functionality from HelloWorld to ElectronIntegration
- Improve UI design and component organization
